+++ crates/ruma-client-api/src/r0/session.rs
@@ -6,4 +6,3 @@ pub mod login_fallback;
 pub mod logout;
 pub mod logout_all;
 pub mod sso_login;
-pub mod sso_login_with_provider;