+++ crates/ruma-client-api/src/session/sso_login_with_provider.rs
@@ -1,0 +1,3 @@
+//! Get the SSO login identity provider url.
+
+pub mod v3;