+++ crates/ruma-client-api/CHANGELOG.md
@@ -42,7 +42,7 @@ Improvements:
 * Move MSC2858 - Multiple SSO Identity Providers out of the `unstable-pre-spec` feature flag, this
   includes:
   * The `r0::session::get_login_types::{IdentityProvider, IdentityProviderBrand}` types
-  * The `r0::session::sso_login_with_provider` endpoint
+  * The `session::sso_login_with_provider::v3` endpoint
 
 # 0.12.3